refactor: rename source/ directory to src/
