+++ package.json
@@ -5,7 +5,7 @@
     "version": "0.0.0",
     "license": "MIT",
     "scripts": {
-        "clear": "rm -rf node_modules/ source/flowtype/node_modules/ source/typescript/node_modules/ source/flowtype/bundle.js source/typescript/bundle.js",
+        "clear": "rm -rf node_modules/ src/flowtype/node_modules/ src/typescript/node_modules/ src/flowtype/bundle.js src/typescript/bundle.js",
         "test": "jasmine test/specs/**",
         "build": "./tasks/build.sh"
     },
