+++ src/flowtype/index.js
